@@ -5,7 +5,7 @@
   "license": "CC",
   "scripts": {
     "analyze": "ANALYZE=true next build",
-    "dev": "next",
+    "dev": "next-remote-watch ./src/content",
     "build": "next build && node ./scripts/generateRSS.js && node ./scripts/generateRedirects.js && node ./scripts/downloadFonts.js",
     "lint": "next lint",
     "lint:fix": "next lint --fix",
@@ -22,26 +22,29 @@
     "check-all": "npm-run-all prettier lint:fix tsc"
   },
   "dependencies": {
-    "@codesandbox/sandpack-react": "v0.19.8-experimental.4",
+    "@codesandbox/sandpack-react": "1.7.2",
     "@docsearch/css": "3.0.0-alpha.41",
     "@docsearch/react": "3.0.0-alpha.41",
-    "@headlessui/react": "^1.3.0",
-    "@mdx-js/react": "^1.6.16",
+    "@headlessui/react": "^1.7.0",
     "body-scroll-lock": "^3.1.3",
     "classnames": "^2.2.6",
     "date-fns": "^2.16.1",
     "debounce": "^1.2.1",
     "ga-lite": "^2.1.4",
     "github-slugger": "^1.3.0",
-    "next": "12.1.7-canary.11",
+    "next": "12.3.2-canary.7",
+    "next-remote-watch": "^1.0.0",
     "parse-numeric-range": "^1.2.0",
-    "react": "0.0.0-experimental-82c64e1a4-20220520",
-    "react-collapsed": "3.1.0",
-    "react-dom": "0.0.0-experimental-82c64e1a4-20220520",
-    "scroll-into-view-if-needed": "^2.2.25"
+    "react": "0.0.0-experimental-cb5084d1c-20220924",
+    "react-collapsed": "npm:@gaearon/react-collapsed@3.1.0-forked.1",
+    "react-dom": "0.0.0-experimental-cb5084d1c-20220924",
+    "remark-gfm": "^3.0.1"
   },
   "devDependencies": {
-    "@mdx-js/loader": "^1.6.16",
+    "@babel/core": "^7.12.9",
+    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
+    "@babel/preset-react": "^7.18.6",
+    "@mdx-js/mdx": "^2.1.3",
     "@types/body-scroll-lock": "^2.6.1",
     "@types/classnames": "^2.2.10",
     "@types/debounce": "^1.2.1",
@@ -51,8 +54,8 @@
     "@types/parse-numeric-range": "^0.0.1",
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.5",
-    "@typescript-eslint/eslint-plugin": "2.x",
-    "@typescript-eslint/parser": "2.x",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
     "asyncro": "^3.0.0",
     "autoprefixer": "^10.4.2",
     "babel-eslint": "10.x",
@@ -63,7 +66,7 @@
     "eslint-plugin-import": "2.x",
     "eslint-plugin-jsx-a11y": "6.x",
     "eslint-plugin-react": "7.x",
-    "eslint-plugin-react-hooks": "2.x",
+    "eslint-plugin-react-hooks": "experimental",
     "fs-extra": "^9.0.1",
     "globby": "^11.0.1",
     "gray-matter": "^4.0.2",
@@ -71,6 +74,7 @@
     "is-ci": "^3.0.1",
     "lint-staged": ">=10",
     "mdast-util-to-string": "^1.1.0",
+    "metro-cache": "0.72.2",
     "npm-run-all": "^4.1.5",
     "patch-package": "^6.2.2",
     "postcss": "^8.4.5",
