@@ -89,7 +89,7 @@ async function main(paths) {
     import('remark-parse'),
     import('remark-slug'),
   ]);
-  const unified = unifiedMod.default;
+  const unified = unifiedMod.unified;
   const remarkParse = remarkParseMod.default;
   const remarkSlug = remarkSlugMod.default;
   modules = {unified, remarkParse, remarkSlug};
