@@ -2,9 +2,9 @@
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
 
-import * as React from 'react';
+import {memo} from 'react';
 
-export const IconWarning = React.memo<JSX.IntrinsicElements['svg']>(
+export const IconWarning = memo<JSX.IntrinsicElements['svg']>(
   function IconWarning({className}) {
     return (
       <svg
@@ -24,5 +24,3 @@ export const IconWarning = React.memo<JSX.IntrinsicElements['svg']>(
     );
   }
 );
-
-IconWarning.displayName = 'IconWarning';
