@@ -3,20 +3,17 @@
  */
 
 import cx from 'classnames';
-import * as React from 'react';
 import {useTocHighlight} from './useTocHighlight';
+import type {Toc} from '../MDX/TocContext';
 
-export function Toc({
-  headings,
-}: {
-  headings: Array<{url: string; text: React.ReactNode; depth: number}>;
-}) {
+export function Toc({headings}: {headings: Toc}) {
   const {currentIndex} = useTocHighlight();
   // TODO: We currently have a mismatch between the headings in the document
   // and the headings we find in MarkdownPage (i.e. we don't find Recap or Challenges).
   // Select the max TOC item we have here for now, but remove this after the fix.
   const selectedIndex = Math.min(currentIndex, headings.length - 1);
   return (
+<<<<<<< HEAD
     <nav
       role="navigation"
       className="pt-6 fixed top-10 right-0"
@@ -28,10 +25,17 @@ export function Toc({
       <h2 className="mb-3 lg:mb-3 uppercase tracking-wide font-bold text-sm text-secondary dark:text-secondary-dark px-4 w-full">
         목차
       </h2>
+=======
+    <nav role="navigation" className="pt-[22px] sticky top-0 right-0">
+      {headings.length > 0 && (
+        <h2 className="mb-3 lg:mb-3 uppercase tracking-wide font-bold text-sm text-secondary dark:text-secondary-dark px-4 w-full">
+          On this page
+        </h2>
+      )}
+>>>>>>> e3073b03a5b9eff4ef12998841b9e56120f37e26
       <div className="h-full overflow-y-auto pl-4 max-h-[calc(100vh-7.5rem)]">
         <ul className="space-y-2 pb-16">
-          {headings &&
-            headings.length > 0 &&
+          {headings.length > 0 &&
             headings.map((h, i) => {
               if (h.url == null) {
                 // TODO: only log in DEV
