@@ -2,8 +2,7 @@
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
 
-import React from 'react';
-import {siteConfig} from 'siteConfig';
+import {useState, useRef, useEffect} from 'react';
 
 const TOP_OFFSET = 75;
 
@@ -24,10 +23,10 @@ export function getHeaderAnchors(): HTMLAnchorElement[] {
  * Sets up Table of Contents highlighting.
  */
 export function useTocHighlight() {
-  const [currentIndex, setCurrentIndex] = React.useState<number>(0);
-  const timeoutRef = React.useRef<number | null>(null);
+  const [currentIndex, setCurrentIndex] = useState<number>(0);
+  const timeoutRef = useRef<number | null>(null);
 
-  React.useEffect(() => {
+  useEffect(() => {
     function updateActiveLink() {
       const pageHeight = document.body.scrollHeight;
       const scrollPosition = window.scrollY + window.innerHeight;
