@@ -1,21 +1,27 @@
 /*
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
-import cn from 'classnames';
-import * as React from 'react';
-const CodeBlock = React.lazy(() => import('./CodeBlock'));
 
-export default React.memo(function CodeBlockWrapper(props: {
+import * as React from 'react';
+import cn from 'classnames';
+import {lazy, memo, Suspense} from 'react';
+const CodeBlock = lazy(() => import('./CodeBlock'));
+
+export default memo(function CodeBlockWrapper(props: {
+  children: React.ReactNode & {
+    props: {
+      className: string;
+      children: string;
+      meta?: string;
+    };
+  };
   isFromPackageImport: boolean;
-  children: string;
-  className?: string;
-  metastring: string;
   noMargin?: boolean;
   noMarkers?: boolean;
 }): any {
   const {children, isFromPackageImport} = props;
   return (
-    <React.Suspense
+    <Suspense
       fallback={
         <pre
           className={cn(
@@ -28,6 +34,6 @@ export default React.memo(function CodeBlockWrapper(props: {
         </pre>
       }>
       <CodeBlock {...props} />
-    </React.Suspense>
+    </Suspense>
   );
 });
