@@ -2,12 +2,13 @@
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
 
+import {isValidElement} from 'react';
 import * as React from 'react';
 import cn from 'classnames';
 import {IconWarning} from '../Icon/IconWarning';
 import {IconError} from '../Icon/IconError';
 
-type LogLevel = 'info' | 'warning' | 'error';
+type LogLevel = 'warning' | 'error' | 'info';
 
 interface ConsoleBlockProps {
   level?: LogLevel;
@@ -27,16 +28,12 @@ const Box = ({
 }) => (
   <div className={className} style={{width, height, ...customStyles}}></div>
 );
-Box.displayName = 'Box';
 
-function ConsoleBlock({level = 'info', children}: ConsoleBlockProps) {
-  let message: string | undefined;
+function ConsoleBlock({level = 'error', children}: ConsoleBlockProps) {
+  let message: React.ReactNode | null;
   if (typeof children === 'string') {
     message = children;
-  } else if (
-    React.isValidElement(children) &&
-    typeof children.props.children === 'string'
-  ) {
+  } else if (isValidElement(children)) {
     message = children.props.children;
   }
 
