@@ -51,8 +51,6 @@ const Heading = forwardRefWithAs<HeadingProps, 'div'>(function Heading(
   );
 });
 
-Heading.displayName = 'Heading';
-
 export const H1 = ({className, ...props}: HeadingProps) => (
   <Heading
     as="h1"
@@ -71,6 +69,7 @@ export const H2 = ({className, ...props}: HeadingProps) => (
     {...props}
   />
 );
+
 export const H3 = ({className, ...props}: HeadingProps) => (
   <Heading
     as="h3"
