@@ -2,7 +2,7 @@
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
 
-import * as React from 'react';
+import {Children, cloneElement} from 'react';
 import NextLink from 'next/link';
 import cn from 'classnames';
 
@@ -16,16 +16,14 @@ function Link({
 }: JSX.IntrinsicElements['a']) {
   const classes =
     'inline text-link dark:text-link-dark break-normal border-b border-link border-opacity-0 hover:border-opacity-100 duration-100 ease-in transition leading-normal';
-  const modifiedChildren = React.Children.toArray(children).map(
-    (child: any) => {
-      if (child.props?.mdxType && child.props?.mdxType === 'inlineCode') {
-        return React.cloneElement(child, {
-          isLink: true,
-        });
-      }
-      return child;
+  const modifiedChildren = Children.toArray(children).map((child: any) => {
+    if (child.type?.mdxName && child.type?.mdxName === 'inlineCode') {
+      return cloneElement(child, {
+        isLink: true,
+      });
     }
-  );
+    return child;
+  });
 
   if (!href) {
     // eslint-disable-next-line jsx-a11y/anchor-has-content
@@ -53,7 +51,5 @@ function Link({
     </>
   );
 }
-
-Link.displayName = 'Link';
 
 export default Link;
