@@ -1,57 +1,81 @@
 /*
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
-import React from 'react';
-// @ts-ignore
+import {memo, useRef, useState} from 'react';
 import {flushSync} from 'react-dom';
 import {
   useSandpack,
   useActiveCode,
   SandpackCodeEditor,
-  SandpackThemeProvider,
-  SandpackReactDevTools,
+  // SandpackReactDevTools,
+  SandpackLayout,
 } from '@codesandbox/sandpack-react';
-import scrollIntoView from 'scroll-into-view-if-needed';
 import cn from 'classnames';
 
 import {IconChevron} from 'components/Icon/IconChevron';
 import {NavigationBar} from './NavigationBar';
 import {Preview} from './Preview';
-import {CustomTheme} from './Themes';
+
 import {useSandpackLint} from './useSandpackLint';
 
-// Workaround for https://github.com/reactjs/reactjs.org/issues/4686#issuecomment-1137402613.
-const emptyArray: Array<any> = [];
-
-export function CustomPreset({
-  isSingleFile,
+export const CustomPreset = memo(function CustomPreset({
   showDevTools,
   onDevToolsLoad,
   devToolsLoaded,
+  providedFiles,
 }: {
-  isSingleFile: boolean;
   showDevTools: boolean;
   devToolsLoaded: boolean;
   onDevToolsLoad: () => void;
+  providedFiles: Array<string>;
 }) {
   const {lintErrors, lintExtensions} = useSandpackLint();
-  const lineCountRef = React.useRef<{[key: string]: number}>({});
-  const containerRef = React.useRef<HTMLDivElement>(null);
   const {sandpack} = useSandpack();
   const {code} = useActiveCode();
-  const [isExpanded, setIsExpanded] = React.useState(false);
-
-  const {activePath} = sandpack;
-  if (!lineCountRef.current[activePath]) {
-    lineCountRef.current[activePath] = code.split('\n').length;
+  const {activeFile} = sandpack;
+  const lineCountRef = useRef<{[key: string]: number}>({});
+  if (!lineCountRef.current[activeFile]) {
+    lineCountRef.current[activeFile] = code.split('\n').length;
   }
-  const lineCount = lineCountRef.current[activePath];
-  const isExpandable = lineCount > 16 || isExpanded;
+  const lineCount = lineCountRef.current[activeFile];
+  const isExpandable = lineCount > 16;
+  return (
+    <SandboxShell
+      showDevTools={showDevTools}
+      onDevToolsLoad={onDevToolsLoad}
+      devToolsLoaded={devToolsLoaded}
+      providedFiles={providedFiles}
+      lintErrors={lintErrors}
+      lintExtensions={lintExtensions}
+      isExpandable={isExpandable}
+    />
+  );
+});
 
+const SandboxShell = memo(function SandboxShell({
+  showDevTools,
+  onDevToolsLoad,
+  devToolsLoaded,
+  providedFiles,
+  lintErrors,
+  lintExtensions,
+  isExpandable,
+}: {
+  showDevTools: boolean;
+  devToolsLoaded: boolean;
+  onDevToolsLoad: () => void;
+  providedFiles: Array<string>;
+  lintErrors: Array<any>;
+  lintExtensions: Array<any>;
+  isExpandable: boolean;
+}) {
+  const containerRef = useRef<HTMLDivElement>(null);
+  const [isExpanded, setIsExpanded] = useState(false);
   return (
     <>
       <div
         className="shadow-lg dark:shadow-lg-dark rounded-lg"
+<<<<<<< HEAD
         ref={containerRef}
       >
         <NavigationBar showDownload={isSingleFile} />
@@ -89,10 +113,43 @@ export function CustomPreset({
                   if (!nextIsExpanded && containerRef.current !== null) {
                     scrollIntoView(containerRef.current, {
                       scrollMode: 'if-needed',
+=======
+        ref={containerRef}>
+        <NavigationBar providedFiles={providedFiles} />
+        <SandpackLayout
+          className={cn(
+            showDevTools && devToolsLoaded && 'sp-layout-devtools',
+            !(isExpandable || isExpanded) && 'rounded-b-lg overflow-hidden',
+            isExpanded && 'sp-layout-expanded'
+          )}>
+          <Editor lintExtensions={lintExtensions} />
+          <Preview
+            className="order-last xl:order-2"
+            isExpanded={isExpanded}
+            lintErrors={lintErrors}
+          />
+          {(isExpandable || isExpanded) && (
+            <button
+              translate="yes"
+              className="sandpack-expand flex text-base justify-between dark:border-card-dark bg-wash dark:bg-card-dark items-center z-10 p-1 w-full order-2 xl:order-last border-b-1 relative top-0"
+              onClick={() => {
+                const nextIsExpanded = !isExpanded;
+                flushSync(() => {
+                  setIsExpanded(nextIsExpanded);
+                });
+                if (!nextIsExpanded && containerRef.current !== null) {
+                  // @ts-ignore
+                  if (containerRef.current.scrollIntoViewIfNeeded) {
+                    // @ts-ignore
+                    containerRef.current.scrollIntoViewIfNeeded();
+                  } else {
+                    containerRef.current.scrollIntoView({
+>>>>>>> e3073b03a5b9eff4ef12998841b9e56120f37e26
                       block: 'nearest',
                       inline: 'nearest',
                     });
                   }
+<<<<<<< HEAD
                 }}
               >
                 <span className="flex p-2 focus:outline-none text-primary dark:text-primary-dark">
@@ -108,9 +165,42 @@ export function CustomPreset({
 
           {showDevTools && (
             <SandpackReactDevTools onLoadModule={onDevToolsLoad} />
+=======
+                }
+              }}>
+              <span className="flex p-2 focus:outline-none text-primary dark:text-primary-dark">
+                <IconChevron
+                  className="inline mr-1.5 text-xl"
+                  displayDirection={isExpanded ? 'up' : 'down'}
+                />
+                {isExpanded ? 'Show less' : 'Show more'}
+              </span>
+            </button>
+>>>>>>> e3073b03a5b9eff4ef12998841b9e56120f37e26
           )}
-        </SandpackThemeProvider>
+        </SandpackLayout>
+
+        {/* {showDevTools && (
+          // @ts-ignore TODO(@danilowoz): support devtools
+          <SandpackReactDevTools onLoadModule={onDevToolsLoad} />
+        )} */}
       </div>
     </>
   );
-}
+});
+
+const Editor = memo(function Editor({
+  lintExtensions,
+}: {
+  lintExtensions: Array<any>;
+}) {
+  return (
+    <SandpackCodeEditor
+      showLineNumbers
+      showInlineErrors
+      showTabs={false}
+      showRunButton={false}
+      extensions={lintExtensions}
+    />
+  );
+});
