@@ -8,7 +8,7 @@ export interface ResetButtonProps {
   onReset: () => void;
 }
 
-export const ResetButton: React.FC<ResetButtonProps> = ({onReset}) => {
+export function ResetButton({onReset}: ResetButtonProps) {
   return (
     <button
       className="text-sm text-primary dark:text-primary-dark inline-flex items-center hover:text-link duration-100 ease-in transition mx-1"
@@ -18,4 +18,4 @@ export const ResetButton: React.FC<ResetButtonProps> = ({onReset}) => {
       <IconRestart className="inline ml-1 mr-1 relative" /> Reset
     </button>
   );
-};
+}
