@@ -2,7 +2,6 @@
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
 
-import * as React from 'react';
 import Breadcrumbs from 'components/Breadcrumbs';
 import Tag from 'components/Tag';
 import {RouteTag} from './Layout/useRouteMeta';
@@ -22,7 +21,7 @@ function PageHeading({
   tags = [],
 }: PageHeadingProps) {
   return (
-    <div className="px-5 sm:px-12 pt-5">
+    <div className="px-5 sm:px-12 pt-8 sm:pt-7 lg:pt-5">
       <div className="max-w-4xl ml-0 2xl:mx-auto">
         {tags ? <Breadcrumbs /> : null}
         <H1 className="mt-0 text-primary dark:text-primary-dark -mx-.5 break-words">
