@@ -2,19 +2,15 @@
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
 
-import * as React from 'react';
+import {useEffect} from 'react';
 import {AppProps} from 'next/app';
 import {useRouter} from 'next/router';
 import {ga} from '../utils/analytics';
+
 import '@docsearch/css';
 import '../styles/algolia.css';
 import '../styles/index.css';
 import '../styles/sandpack.css';
-import '@codesandbox/sandpack-react/dist/index.css';
-
-const EmptyAppShell = ({children}: {children: React.ReactNode}) => (
-  <>{children}</>
-);
 
 if (typeof window !== 'undefined') {
   if (process.env.NODE_ENV === 'production') {
@@ -28,7 +24,23 @@ if (typeof window !== 'undefined') {
 
 export default function MyApp({Component, pageProps}: AppProps) {
   const router = useRouter();
-  React.useEffect(() => {
+
+  useEffect(() => {
+    // Taken from StackOverflow. Trying to detect both Safari desktop and mobile.
+    const isSafari = /^((?!chrome|android).)*safari/i.test(navigator.userAgent);
+    if (isSafari) {
+      // This is kind of a lie.
+      // We still rely on the manual Next.js scrollRestoration logic.
+      // However, we *also* don't want Safari grey screen during the back swipe gesture.
+      // Seems like it doesn't hurt to enable auto restore *and* Next.js logic at the same time.
+      history.scrollRestoration = 'auto';
+    } else {
+      // For other browsers, let Next.js set scrollRestoration to 'manual'.
+      // It seems to work better for Chrome and Firefox which don't animate the back swipe.
+    }
+  }, []);
+
+  useEffect(() => {
     const handleRouteChange = (url: string) => {
       ga('set', 'page', url);
       ga('send', 'pageview');
@@ -39,16 +51,5 @@ export default function MyApp({Component, pageProps}: AppProps) {
     };
   }, [router.events]);
 
-  let AppShell = (Component as any).appShell || EmptyAppShell;
-  // In order to make sidebar scrolling between pages work as expected
-  // we need to access the underlying MDX component.
-  if ((Component as any).isMDXComponent) {
-    AppShell = (Component as any)({}).props.originalType.appShell;
-  }
-
-  return (
-    <AppShell>
-      <Component {...pageProps} />
-    </AppShell>
-  );
+  return <Component {...pageProps} />;
 }
